+++ app/src/main/java/com/segg3/coursemanager/shared/models/UsersViewModel.java
@@ -3,6 +3,7 @@ package com.segg3.coursemanager.shared.models;
 
 import android.util.Log;
 
+import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
@@ -14,22 +15,23 @@ import com.segg3.coursemanager.Admin;
 import com.segg3.coursemanager.Instructor;
 import com.segg3.coursemanager.Student;
 import com.segg3.coursemanager.User;
+import com.segg3.coursemanager.shared.BooleanCallback;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 
 public class UsersViewModel extends ViewModel {
-    private MutableLiveData<HashMap<String, User>> users_impl;
-    private MutableLiveData<List<User>> users;
     private static final CollectionReference userDB = FirebaseFirestore.getInstance().collection("Users");
-
     private static final String PASSWORD_KEY = "password";
     private static final String TYPE_KEY = "type";
     private static final String USERNAME_KEY = "userName";
+    private MutableLiveData<HashMap<String, User>> users_impl;
+    private MutableLiveData<List<User>> users;
+    private User loggedUser;
 
-    public MutableLiveData<List<User>> getUsers(){
-        if (users_impl == null){
+    public LiveData<List<User>> getUsers() {
+        if (users_impl == null) {
             users_impl = new MutableLiveData<>();
             users_impl.setValue(new HashMap<>());
             users = new MutableLiveData<>();
@@ -39,14 +41,40 @@ public class UsersViewModel extends ViewModel {
         return users;
     }
 
-    public void loadUsers(){
+    public User getLoggedUser() {
+        return loggedUser;
+    }
+
+    public void logOut() {
+        loggedUser = null;
+    }
+
+    public BooleanCallback loginUser(String name, String password) {
+        BooleanCallback callback = new BooleanCallback();
+        userDB.whereEqualTo(PASSWORD_KEY, password).whereEqualTo(USERNAME_KEY, name).get().addOnCompleteListener(task -> {
+            if (!task.isSuccessful() || task.getResult().size() == 0) {
+                Log.i("FIRE", "loginUser: tried login but, task was " + task.isSuccessful() + " with " + task.getResult().size() + " matches");
+                callback.set(false);
+            } else {
+                QueryDocumentSnapshot userDoc = task.getResult().iterator().next();
+                User u = deserializeUser(userDoc);
+                loggedUser = u;
+                Log.d("FIRE", "Setting logged in user as " + u.toString());
+                callback.set(true);
+            }
+        });
+
+        return callback;
+    }
+
+    public void loadUsers() {
         userDB.addSnapshotListener((snapshots, e) -> {
             if (e != null)
                 return;
             for (DocumentChange dc : snapshots.getDocumentChanges()) {
                 QueryDocumentSnapshot doc = dc.getDocument();
                 User userTemp = deserializeUser(doc);
-                if (userTemp == null){
+                if (userTemp == null) {
                     Log.w("FIRE", "Object: " + doc.getId() + " doesn't match user signature");
                     continue;
                 }
@@ -66,11 +94,15 @@ public class UsersViewModel extends ViewModel {
         });
     }
 
-    public void deleteUser(int position) {
+    public boolean deleteUser(int position) {
+        if (position < 0 || position >= users.getValue().size())
+            return false;
+
         userDB.document(users.getValue().get(position).getUserID()).delete();
+        return true;
     }
 
-    private HashMap<String, Object> serializeUser(String name, String password, String type){
+    private HashMap<String, Object> serializeUser(String name, String password, String type) {
         HashMap<String, Object> serialized = new HashMap<>();
         serialized.put(USERNAME_KEY, name);
         serialized.put(PASSWORD_KEY, password);
@@ -78,9 +110,9 @@ public class UsersViewModel extends ViewModel {
         return serialized;
     }
 
-    private User deserializeUser(QueryDocumentSnapshot doc){
+    private User deserializeUser(QueryDocumentSnapshot doc) {
         User user = null;
-        switch (doc.get(TYPE_KEY, String.class).toLowerCase()){
+        switch (doc.get(TYPE_KEY, String.class).toLowerCase()) {
             case "admin":
                 user = new Admin(doc.getId(), doc.get(USERNAME_KEY, String.class), "", doc.get(USERNAME_KEY, String.class), "");
                 break;
@@ -93,22 +125,45 @@ public class UsersViewModel extends ViewModel {
             default:
                 break;
         }
+        if (user == null) {
+            Log.d("FIRE", "deserializeUser: Error deserializing user " + doc.getId() + " of type " + doc.get(TYPE_KEY, String.class));
+        } else {
+            Log.d("FIRE", "deserializeUser: Deserialized user " + user.toString());
+        }
         return user;
     }
 
-    public void addUser(String name, String password, String type) {
+    public boolean addUser(String name, String password, String type) {
+        for (User u : users.getValue())
+            if (u.name.equals(name))
+                return false;
+        // Add user ok
         userDB.add(serializeUser(name, password, type));
+        return true;
     }
 
-    public void editUser(int position, String name, String password, String type) {
+    public boolean editUser(int position, String name, String password, String type) {
+        if (position < 0 || position >= users.getValue().size())
+            return false;
+
+        String originalName = users.getValue().get(position).getUsername();
         HashMap<String, Object> serialized = serializeUser(name, password, type);
         if (password == null || password.isEmpty())
             serialized.remove(PASSWORD_KEY); // Keep the password
         if (type == null || type.isEmpty())
             serialized.remove(TYPE_KEY); // Keep the user type
+        if (name == null || name.isEmpty()) {
+            // Check for repeated course codes
+            for (User c : users.getValue())
+                if (c.getUsername().equals(name) && !c.getUsername().equals(originalName))
+                    return false;
+            serialized.remove(USERNAME_KEY); // Keep the user name
+        }
+
 
         String id = users.getValue().get(position).getUserID();
         userDB.document(id).update(serialized);
+        return true;
     }
 
 }
